co-locate unit tests with their target source
