+++ frontend/src/utilities/__tests__/notebookControllerUtils.spec.ts
